Adding proposed changes to the config system
These changes allow the user to load AWS keys from ~/.aws/credentials, which
is the preferred way as it prevents having to put keys in code. The user can
do nothing, in which case the default profile is loaded, or they can specify
the `credentials` as a string, which will set the profile to use.

The old system of manually specifying keys has been kept, as it may be useful.